Create README.md
THE STRUGGLE OF W.I.N.

This developed struggle cannot be only on the basis of wanting to be in power and ruling, on the basis of wanting rank and name, on the basis of wanting to uphold the nation, on the basis of wanting to eradicate tyranny (even if ordered) and others.

If we fight and jihad for one of the above purposes, our struggle is in vain. Not yet classified as fisabilillah jihad.

If we fight on the basis of nationalism only, Rasulullah SAW reminds us in a Hadith which means, "There is no national ideology in the teachings of Islam".

The national ideology is a narrow ideology that will invite people to fight, because each has a race. Such a struggle will only provoke factions.

Then Rasulullah said again, "Not from our group, people who fight for nationalism".

In fact, if we fight to uphold the religion, the nation will defend itself. But if we fight to develop the nation, religion will be neglected. Until then, the nation cannot necessarily be developed, religion has been shattered. Most likely the national ideology in the struggle will take over the ideology of jihad because of religion.

The same goes for the struggle to uphold religion through the concept of capitalism, socialism, open democratic system@elections and so on. Everything is an inappropriate struggle to be judged as fisabilillah.

This is what we need to watch out for. Although in the struggle, the verses of the Quran and Hadith are like fire coming out, but because the struggle is not accurate, people who fight like this can be said to have used the Quran and Hadith to achieve their goals. And for us, too, should not be affected by such a struggle. It is not classified as fisabilillah jihad !.

The real fisabilillah jihad is a jihad in which every inch of the struggle is in accordance with the will of GOD, namely the will of the Quran and Sunnah.

The Quran and the Sunnah must be implemented at all levels. Not only in the country but also in oneself, in the household, in society, in education, in the economy, in politics, and so on to the question of the universe.

Then the implementation must also be correct until the birth and inner law is taken care of properly. If everything is correct according to the requirements of the Quran and Sunnah, that is what is called fisabilillah jihad, and if the person who fought is killed or killed, God willing, his value is as a martyr of GOD. Only then is the mass maintained by ALLAH SWT.

Unfortunately, there are among the Muslim fighters at the end of this age, a kind of true joy fighting to uphold the religion of GOD in the country but oneself is neglected, the wife's son is still exposing his aurat, he is involved in immorality and ugliness. Fighters